pyproject.toml
[project]
name = "ell715-assignment-5"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "joblib>=1.5.2",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.7",
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "pytest>=9.0.1",
    "scikit-image>=0.25.2",
    "scipy>=1.15.3",
    "seaborn>=0.13.2",
    "tqdm>=4.67.1",
]
